split up implementation
